V6: 130-feature stack + Colab one-command training pipeline
New files:
- colab_v6_train.py     -- single Colab cell: download + compute V6 + train on GPU
- features_v6.py        -- 36 new features: Donchian, AVWAP, VP, Sweeps, SMA, OF, Fib, BB
- train_v6.py           -- local training pipeline (V6_full 130f + V6_micro 36f)
- download_v6_ohlcv.py  -- local OHLCV downloader (fallback)
- blind_backtest.py     -- OOS backtest with V5+V6 70/30 blending
- live_demo_engine.py   -- live engine with V6 auto-detection + blending
- colab_unified_pipeline.py -- updated pipeline with inline V6 builder

Updated:
- README.md             -- quick start: git clone + one Colab command
- .gitignore            -- exclude data/, model weights, backups, API keys
- requirements.txt      -- clean dependency list
- .env.example          -- API key template
- data/README.md        -- instructions for populating model weights
- features_v4/v5        -- compatibility with V6 stack
- trading_config.py     -- LOOSEN_MODE=False, ALLOW_HALT_REGIMES=False

Architecture:
  p_final = 0.30 * p_v5 + 0.70 * p_v6_full
  Gate:     p_v6_micro >= 0.50 required

+++ features_v4.py
@@ -191,7 +191,7 @@ def build_v4_extras(close, high, low, volume, ofi, taker_buy_volume,
     interactions; if None, BTC features are zeroed.
     """
     n = len(close)
-    F = np.zeros((n, len(FEATURE_NAMES_V4_EXTRA)), dtype=np.float64)
+    F = np.zeros((n, len(FEATURE_NAMES_V4_EXTRA)), dtype=np.float32)
 
     # log returns at 1m for downstream rolling stats
     log_ret = np.zeros(n)
@@ -317,4 +317,4 @@ def build_v4_from_components(F_v2, close, high, low, volume, ofi,
     F_v3 = build_v3_from_v2(F_v2, close, high, low, atr_arr)             # (n, 29)
     F_extras = build_v4_extras(close, high, low, volume, ofi,
                                taker_buy_volume, timestamps_ms, btc_close)  # (n, 33)
-    return np.hstack([F_v3, F_extras])
+    return np.hstack([F_v3, F_extras]).astype(np.float32)
